feat: Implement UserRegistrationController with Enhanced Validation and Refactoring
- Developed `UserRegistrationController` to handle user registration, updates, retrieval, and deletion.
- Added `IUserRegistrationService` interface to promote loose coupling, enhance testability, and improve flexibility.
- Applied the `@Valid` annotation to enforce request validation.
- Implemented the `buildResponse` method to standardize API responses.
- Added service logic to manage user registration CRUD operations.
- Introduced `convertToDto` for efficient conversion between `UserEntity` and `UserDto`.
- Refactored `UserRegistrationRequest` and `UserRegistrationResponse` into a dedicated folder for request and response handling.
- Renamed `UserRegistrationResponse` to `UserDto` for improved clarity and consistency.

Closes #70
