+++ .idea/jpb-settings.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="DatabaseMigrationSettings" lastSelectedDirectory="user-service/src/main/resources" />
+  <component name="DatabaseMigrationSettings" lastSelectedDirectory="apps/backend/user-service/src/main/java/com/bytebandit/userservice/repository" />
 </project>
