+++ apps/backend/user-service/src/main/java/com/bytebandit/userservice/repository/UserRepository.java
@@ -24,3 +24,4 @@ public interface UserRepository extends JpaRepository<UserEntity, UUID> {
         @Param("expiresAt") java.sql.Timestamp expiresAt
     );
 }
+
