+++ apps/backend/user-service/src/main/java/com/bytebandit/userservice/request/UserRegistrationRequest.java
@@ -4,10 +4,10 @@ import com.bytebandit.userservice.annotation.ValidEmail;
 import com.bytebandit.userservice.annotation.ValidPassword;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Builder;
-import lombok.AllArgsConstructor;
 
 @Data
 @Builder
