+++ apps/backend/user-service/src/main/resources/schema.sql
@@ -66,7 +66,3 @@ BEGIN
 END;
 $$ LANGUAGE plpgsql;
 
-
-
-
-
